Gets up to 3 save profiles working with menu
I know have the menu for loading and starting a new game up and running!

There is still a little bit of weirdness if you load a profile that isnt
already saved, with
some levels already unlocked on another profile, it will unlock that
profiles levels too.
Need to fix that, and clean the code up drastically. Putting up for
testing.